+++ src/feature/pageTemplate/Banner/index.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import * as S from './style';
-import Avatar from '../../components/Avatar';
+import Avatar from '../../../components/Avatar';
 // 이모션 변수 사용
 // 현재 문제는 윈도우에서 좌표값이 잡힘 배너는 위에만 박혀있어서 좌표값이 바뀌지 않음
 const Banner = ({ isScrollDown = false }) => {
