+++ src/feature/pageTemplate/Banner/style.jsx
