+++ src/feature/pageTemplate/Header/index.jsx
@@ -1,10 +1,9 @@
 // import PropTypes from 'prop-types';
 import PropTypes from 'prop-types';
-import Avatar from '../../components/Avatar';
-import Icon from '../../components/Icon';
+import Avatar from '../../../components/Avatar';
+import Icon from '../../../components/Icon';
 import * as S from './style';
-import iconAlarm from '../../assets/icon_alarm_line.svg';
-import Badge from '../../components/Badge';
+import Badge from '../../../components/Badge';
 
 // 따봉왕, 코인왕 아바타 이미지 props로 받아야함
 
@@ -22,7 +21,7 @@ const Header = ({ isScrollDown = false, isAlarm = false }) => {
       <S.PlacedLogo />
       <S.IconWrapper>
         <Badge dot={isAlarm}>
-          <Icon src={iconAlarm} />
+          <Icon name="alarmLine" />
         </Badge>
       </S.IconWrapper>
     </S.Header>
